Rename UserProvider.tsx to hooks.ts
